Added fields which allow adding mobileAppName and link for tickets
git-svn-id: svn://127.0.0.1/flaskdump/trunk@2466 386f392f-2c41-1343-9bf7-4b6fedb51b34

+++ liferay-plugins-sdk-6.2/portlets/flask-user-tailgate-portlet/docroot/WEB-INF/service/com/rumbasolutions/flask/service/TailgateInfoService.java
@@ -69,7 +69,7 @@ public interface TailgateInfoService extends BaseService, InvokableService {
 		java.lang.String tailgateName, java.lang.String tailgateDescription,
 		long eventId, java.lang.String eventName, java.util.Date tailgateDate,
 		long startTime, long endTime, java.lang.String venmoAccountId,
-		long amountToPay, long logoId,
+		double amountToPay, long logoId,
 		com.liferay.portal.service.ServiceContext serviceContext);
 
 	@com.liferay.portal.security.ac.AccessControlled(guestAccessEnabled = true)
@@ -104,7 +104,7 @@ public interface TailgateInfoService extends BaseService, InvokableService {
 		java.lang.String tailgateDescription, long eventId,
 		java.lang.String eventName, java.util.Date tailgateDate,
 		long startTime, long endTime, java.lang.String venmoAccountId,
-		long amountToPay, long logoId,
+		double amountToPay, long logoId,
 		com.liferay.portal.service.ServiceContext serviceContext);
 
 	public void deleteTailgateInfo(long tailgateId,
